anlzr: replace AverageComplexity field by a ComplexityMetric structure
A simple number is not enough for analyzing the complexity of a whole
project. Multiple metrics can be used and can be applied to different
part of a project (per file, per function, per class, etc.). Therefore,
a ComplexityMetric structure allows more flexibility for further
improvements and make a clear separation in the Result structure.