src: add a FIXME and TODO annotations
For the FIXME: In a function (or method), return statements may occur at
multiple locations (eg. into a If statement) and not only at the end of
the function. Therefore, it makes sense to take all of these statements
into account in the API.

For the TODO: Loop statement is more generic term than "For" or "While".
It makes more sense here to use the more abstract form. Moreover, we
also need to improve this Statement by adding the loop condition.