src: add a function for unmarshalling a JSON representation of a Project
The JSON decoder of the standard library (json.Unmarshal) is unable to
unmarshal correctly an interface. Therefore, all Statements are
interpreted as a map[string]interface{} instead of unmarshalling into a
IfStatement, ForStatement, AssignStatement, CallStatelement or
OtherStatement. This method uses json.Unmarshal to unmarshal the JSON,
and then fixeds all Statements by replacing map[string]interface{} by
the appropriate type.

This is not an ideal solution, but for now it's enough. A better
approach would be to write a specific JSON decoder. This will be
probably done later on.