src: remove source file type
In the SourceFile type, there is a field for the Type of source file.
This field can take the value header or source file, but in the end,
this is something very specific to some languages such as C/C++,
Pascal and few others. But this is something very very specific.
Moreover, it does not bring any useful information and generally, it
can be included in a single instance of a SourceFile. So, in order
to simplify the API and to be more "abstract", this commit removes
this field.