src: :warning: break compat' / rename ForStatement into LoopStatement
A "For" statement is something very specific to some languages. Besides,
languages often have multiple loop statements (for, while, do...wile,
repeat, etc.), which is obviously too specific for an abstract
representation. Thus, it makes sense to only have a generic Loop
statement for all kinds of loop and "Loop" instead of "For" is a much
better name then.

Note: The LoopStatement is clearly incomplete and needs to containt more
information about a loop, such as the kind of loop, the stop condition,
etc.