repo: :warning: break compat' / move repo.Repo into src.Repo
The src package defines the API for representing a project and its
source codes so it makes sense to have the Repo type there and not
in a separate pacakge repo. Besides, the repo package has nothing to do
here, it should be moved into the srctool project (a separate project
that provides a CLI tool for adding repository related information to
the JSON produced by the parsers).