src: :warning: break compat' / fix the project unmarshalling
The src.Unmarshal method was not working properly due to several
mistakes. This commit fix them all. The error was not detected because
of an error in the test anlzr/anlzr_test.go: it was still using
json.Unmarshal for unmarshalling go.json insteal of src.Unmarshal.

To summarize what this commit does, here is the list of changes:

* Fix the src unmarshalling
* Fix the test
* Update the test data because the fix for the src.Unmarshal function
  slightly modifies the API, which breaks the test data. The expected
  values of the tests also must be updated because there were, of
  course, some changes in the github.com/golang/go repository since the
  time.