src: add debug info to all errors in the package
Add debug info that include the file name and the line to all error in
the src/ package. In the end, it produces the whole call trace before
the error is encountered.

This commit also updates the test data and test values with the latest
changes in the github.com/golang/go repository.