src: :warning: break compat' / complete the src API
Add several new structures to the src API and modifies existing ones in
order to improve and cover more situations.

The following types are notably created/changed

* Expr (new)
* ExprStmt (new)
* IfStmt
* LoopStmt
* CallExpr (renamed)
* AssignStmt
* DeclStmt (new)
* SwitchStmt (new)
* CaseClause (new)
* StructType (renamed)
* Field

Until the src API is stabilized, the test are disabled because the test
data are not up to date and it won't be able to parse them.