srcanlzr: support reading from stdin + improve file reader
If no argument is supplied when issuing the srcanlzr command, it will
now read input from stdin. The purpose of this change is to allow
pipelining with our other tools (srctool and repotool):

   "srctool some/folder | repotool | srcanlzr"

Besides, this commit also improves the way a file is read: instead of
reading the whole file with ioutil.ReadFile, it reads it with a buffer
and the io.Copy function, which should be safer for big files.